requirements.txt
numpy==1.26.1
opencv-python==4.8.1.78
pandas==2.1.2
scikit-learn==1.3.2
pillow==10.3.0
